PPA_CCC/highway_project_1차/requirements.txt
# MetaDrive 강화학습 프로젝트 필요 패키지
# Python 3.9 - 3.11 권장 (3.13은 호환 문제 있음)

# ============================================================
# 핵심 패키지
# ============================================================

# MetaDrive (PGDrive 후속 버전)
# GitHub에서 직접 설치 필요:
# pip install git+https://github.com/metadriverse/metadrive.git
# 또는 PyPI 버전:
metadrive-simulator>=0.4.0

# 강화학습 프레임워크
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# 구버전 gym (MetaDrive 의존성)
gym>=0.26.0

# ============================================================
# 시각화
# ============================================================

matplotlib>=3.5.0
seaborn>=0.12.0
tensorboard>=2.10.0

# ============================================================
# 데이터 처리
# ============================================================

numpy>=1.21.0,<2.0.0  # MetaDrive 호환성
pandas>=1.3.0

# ============================================================
# 유틸리티
# ============================================================

tqdm>=4.62.0
pyyaml>=6.0

# ============================================================
# MetaDrive 의존성 (수동 설치 필요할 수 있음)
# ============================================================

# 3D 렌더링
panda3d>=1.10.0

# 기하학 처리
shapely>=1.8.0

# 이미지 처리
pillow>=9.0.0

# Cython (빌드용)
Cython>=0.29.0

# ============================================================
# 선택적 패키지
# ============================================================

# GPU 사용 시 (PyTorch)
# torch>=2.0.0
# torchvision>=0.15.0

# 추가 시각화
# opencv-python>=4.5.0

# Jupyter 노트북
# jupyter>=1.0.0
# ipykernel>=6.0.0
